# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Outpost VPN API
  description: |
    Open-source WireGuard VPN & Zero-Trust Network Access platform.
    All features available under Apache 2.0 — no enterprise paywalls.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Romashqua Labs
    url: https://github.com/romashqua/outpost
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://vpn.example.com
    description: Production
...
paths:
...
  /api/v1/devices:
    get:
      operationId: listDevices
      summary: List all devices
      tags: [Devices]
      responses:
        "200":
          description: Device list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"

    post:
      operationId: createDevice
      summary: Register a new device
      tags: [Devices]
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "201":
          description: Device registered (IP auto-assigned)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Device already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/devices/my:
    get:
      operationId: listMyDevices
      summary: List devices for the authenticated user
      tags: [Devices]
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User's device list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        "400":
          description: Missing user_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/devices/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"

    get:
      operationId: getDevice
      summary: Get device by ID
      tags: [Devices]
      responses:
        "200":
          description: Device found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      operationId: deleteDevice
      summary: Delete a device
      tags: [Devices]
      responses:
        "204":
          description: Device deleted
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/devices/{id}/approve:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    post:
      operationId: approveDevice
      summary: Approve device for VPN access
      tags: [Devices]
      responses:
        "200":
          description: Device approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: approved
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/devices/{id}/revoke:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    post:
      operationId: revokeDevice
      summary: Revoke device VPN access
      tags: [Devices]
      responses:
        "200":
          description: Device revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: revoked
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Gateways ────────────────────────────────────────────────────────
...
  /api/v1/devices/enroll:
    post:
      operationId: enrollDevice
      summary: Self-service device enrollment
      tags: [Devices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollDeviceRequest"
      responses:
        "201":
          description: Device enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollDeviceResponse"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Device already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: IP allocation or internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Device Config Download ───────────────────────────────────────────
  /api/v1/devices/{id}/config:
    get:
      operationId: downloadDeviceConfig
      summary: Download WireGuard configuration for a device
      tags: [Devices]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: WireGuard configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceConfigResponse"
        "400":
          description: Invalid UUID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Key generation or network error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── S2S Tunnels ──────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via POST /api/v1/auth/login
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: something went wrong

    # ── Auth ────────────────────────────────────────────────────────
...
    Device:
      type: object
      required: [id, user_id, name, wireguard_pubkey, is_approved, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        wireguard_pubkey:
          type: string
          description: WireGuard public key (base64)
        assigned_ip:
          type: string
          description: Auto-assigned VPN IP address
          example: 10.10.0.2
        is_approved:
          type: boolean
        last_handshake:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateDeviceRequest:
      type: object
      required: [name, wireguard_pubkey]
      properties:
        name:
          type: string
          minLength: 1
          example: MacBook Pro
        wireguard_pubkey:
          type: string
          description: WireGuard public key (base64)

    # ── Gateways ────────────────────────────────────────────────────
...
    EnrollDeviceRequest:
      type: object
      required: [name, wireguard_pubkey]
      properties:
        name:
          type: string
          minLength: 1
          example: iPhone 15
        wireguard_pubkey:
          type: string
          description: WireGuard public key (base64)
    EnrollDeviceResponse:
      type: object
      required: [device_id, address, dns, endpoint, server_public_key, allowed_ips, persistent_keepalive]
      properties:
        device_id:
          type: string
          format: uuid
        address:
          type: string
          description: Assigned VPN address with mask
          example: 10.10.0.3/24
        dns:
          type: array
          items:
            type: string
          example: ["1.1.1.1"]
        endpoint:
          type: string
          example: vpn.example.com:51820
        server_public_key:
          type: string
        allowed_ips:
          type: array
          items:
            type: string
          example: ["0.0.0.0/0"]
        persistent_keepalive:
          type: integer
          example: 25
    DeviceConfigResponse:
      type: object
      required: [config, private_key, public_key]
      properties:
        config:
          type: string
          description: WireGuard INI-format configuration
        private_key:
          type: string
          description: WireGuard private key (base64)
        public_key:
          type: string
          description: WireGuard public key (base64)

    # ── S2S Tunnels ──────────────────────────────────────────────────
...
security:
  - BearerAuth: []